added withMap hoc
